[feat] 에러 메시지 출력

+++ marketplace/src/main/java/store/carjava/marketplace/firebase/event/NewCarAlarmEventListener.java
@@ -24,7 +24,7 @@ public class NewCarAlarmEventListener {
 
     @EventListener
     @Transactional
-    public void handleNewCarAlarmEvent(CarSellEvent event) throws FirebaseMessagingException {
+    public void handleNewCarAlarmEvent(CarSellEvent event){
         //입고된 차랑 동일한 모델을 찜한 사용자를 가져온다.
         Set<Long> userIdList = likeRepository.findUserByMarketplaceCar_CarDetails_Model(event.getModel());
 
@@ -33,28 +33,38 @@ public class NewCarAlarmEventListener {
         if(!userIdList.isEmpty()) {
             tokens = fcmTokenRepository.findAllByUserIdIn(userIdList);
 
-            // 토큰들에 알람 요청
-            MulticastMessage message = MulticastMessage.builder()
-                    .setNotification(Notification.builder()
-                            .setTitle("새 중고차 입고")
-                            .setBody("내가 찜한 차와 같은 모델이 입고됐습니다.")
-                            .build())
-                    .putData("url", "https://chajava.store/carDetail/" + event.getId())
-                    .addAllTokens(tokens)
-                    .build();
+            if(!tokens.isEmpty()) {
+                // 토큰들에 알람 요청
+                MulticastMessage message = MulticastMessage.builder()
+                        .setNotification(Notification.builder()
+                                .setTitle("새 중고차 입고")
+                                .setBody("내가 찜한 차와 같은 모델이 입고 됐습니다")
+                                .build())
+                        .putData("url", "https://chajava.store/carDetail/HUS241223011180")
+                        .addAllTokens(tokens)
+                        .build();
 
-            // 만료된 요청이 오는 토큰 삭제
-            BatchResponse response = FirebaseMessaging.getInstance().sendMulticast(message);
-            if (response.getFailureCount() > 0 ){
-                List<SendResponse> responses = response.getResponses();
-                for (int i = 0; i < responses.size(); i++) {
-                    if (!responses.get(i).isSuccessful()){
-                        // 만료 토큰 삭제
-                        fcmTokenRepository.deleteByFcmToken(tokens.get(i));
+                // 만료된 요청이 오는 토큰 삭제
+                try{
+                    BatchResponse response = FirebaseMessaging.getInstance().sendEachForMulticast(message);
+                    if (response.getFailureCount() > 0 ){
+                        List<SendResponse> responses = response.getResponses();
+                        for (int i = 0; i < responses.size(); i++) {
+                            if (!responses.get(i).isSuccessful()){
+                                // 만료 토큰 삭제
+                                fcmTokenRepository.deleteByFcmToken(tokens.get(i));
+                            }
+                        }
                     }
+                    log.info("알람 전송 완료");
+                }catch (Exception e){
+                    log.info("알람 전송 실패 : " + e.getMessage());
                 }
+
             }
-            log.info("알람 전송 완료");
+
+
+
         }
     }
 }
